Phase 4h: extract NervesHub.Devices.Deployments sub-context
Final step of the Devices cleanup. Extracts the device<->deployment-group
functions out of NervesHub.Devices into a new NervesHub.Devices.Deployments
module. Built on Phase 4g.

Moved (a device's relationship to its deployment group + per-deployment device
counts): get_device_firmware_for_delta_generation_by_deployment_group,
matches_deployment_group?, update_deployment_group, clear_deployment_group,
deployment_device_online, up_to_date_count, updating_count,
waiting_for_update_count, updates_disabled_count, in_penalty_box_count,
remove_unmatched_devices_from_deployment_group, plus private version_match?/
tags_match? and the delta-gen @type decls.

The cluster is self-contained (no back-calls into core Devices), so no
Devices-qualification was needed. Six now-unused aliases removed from devices.ex.

Call sites repointed to Deployments across lib and test (incl. ManagedDeployments
itself, which called four of these). No dynamic-dispatch or Mimic wiring for this
cluster (verified).

With this, devices.ex is down from 1927 to ~783 lines, and the Devices context
is split into six focused sub-modules: CACertificates, Pinning, Health,
Certificates, Updates, Deployments (with delta resolution moved to Firmwares and
signing keys to Accounts).

Verification: mix format (bare, exit 0), credo, dialyzer (0 unskipped), full
test suite (1396 passing).
